重构InputWait，异步逻辑改成同步

+++ Unity/Assets/Scripts/Loader/Plugins/Box2DSharp/Box2DSharp/Testbed.Abstractions/TestSettings.cs
@@ -128,6 +128,6 @@ namespace Testbed.Abstractions
         
         public long instanceId;
 
-        public int StepCount;
+        public long StepCount;
     }
 }
